Remove the need for special xcode version during tests (wooga/atlas-build-unity#137)
> Note Original PR: https://github.com/wooga/atlas-build-unity/pull/137

Description
===========

The old tests had to force a testrun with xcode 12 setup. XCode 13
changed it's output and the tests won't be able to cover over multiple
versions. I decided to make the test case simpler by only testing a single
line of the output for the formatting instead of trying to match the whole
output with regex. I also needed to adjust the output handling from xcodebuild
because error lines are not printed to stderr instead of stdout as before with
Xcode12. I simply merge both streams together for a single log file.

Changes
=======

* ![IMPROVE] simplify xcode format check to be version agnostic
* ![FIX] XCode13 error output handling